generate ServicesByName
