+++ src/lib/app/eventTypes.gen.ts
@@ -8,6 +8,7 @@ import {
 	type JsonSchemaToTypeScriptOptions,
 } from '$lib/util/jsonSchemaToTypescript';
 import {schemas} from '$lib/app/schemas';
+import {services} from '$lib/server/services';
 
 const toParamsName = (name: string): string => name + 'Params';
 const toResponseName = (name: string): string => name + 'Response';
@@ -31,6 +32,7 @@ import type {SvelteComponent} from 'svelte';
 import type {Readable} from 'svelte/store';
 
 import type {ApiResult} from '$lib/server/api';
+import type {Service} from '$lib/server/service';
 import type {Community} from '$lib/vocab/community/community';
 import type {Persona, AccountPersona} from '$lib/vocab/persona/persona';
 import type {Membership} from '$lib/vocab/membership/membership';
@@ -111,6 +113,13 @@ export interface Mutations {
 `.trim(),
 		'',
 	)}
+}
+
+export interface ServicesByName {
+	${Array.from(services.values()).reduce((str, service) => {
+		const {name} = service.event;
+		return str + `${name}: Service<${toParamsName(name)}, ${toResponseResultName(name)}>;`.trim();
+	}, '')}
 }
 
 // generated by ${toRootPath(originId)}
