+++ src/lib/app/eventTypes.ts
@@ -4,6 +4,7 @@ import type {SvelteComponent} from 'svelte';
 import type {Readable} from 'svelte/store';
 
 import type {ApiResult} from '$lib/server/api';
+import type {Service} from '$lib/server/service';
 import type {Community} from '$lib/vocab/community/community';
 import type {Persona, AccountPersona} from '$lib/vocab/persona/persona';
 import type {Membership} from '$lib/vocab/membership/membership';
@@ -476,6 +477,38 @@ export interface Mutations {
 	SelectCommunity: (ctx: DispatchContext<SelectCommunityParams, void>) => void;
 	SelectSpace: (ctx: DispatchContext<SelectSpaceParams, void>) => void;
 	ViewSpace: (ctx: DispatchContext<ViewSpaceParams, void>) => void;
+}
+
+export interface ServicesByName {
+	Ping: Service<PingParams, PingResponseResult>;
+	LoginAccount: Service<LoginAccountParams, LoginAccountResponseResult>;
+	LogoutAccount: Service<LogoutAccountParams, LogoutAccountResponseResult>;
+	CreateAccountPersona: Service<CreateAccountPersonaParams, CreateAccountPersonaResponseResult>;
+	ReadPersona: Service<ReadPersonaParams, ReadPersonaResponseResult>;
+	CreateCommunity: Service<CreateCommunityParams, CreateCommunityResponseResult>;
+	CreateMembership: Service<CreateMembershipParams, CreateMembershipResponseResult>;
+	DeleteMembership: Service<DeleteMembershipParams, DeleteMembershipResponseResult>;
+	CreateSpace: Service<CreateSpaceParams, CreateSpaceResponseResult>;
+	CreateEntity: Service<CreateEntityParams, CreateEntityResponseResult>;
+	UpdateEntity: Service<UpdateEntityParams, UpdateEntityResponseResult>;
+	EraseEntity: Service<EraseEntityParams, EraseEntityResponseResult>;
+	DeleteEntities: Service<DeleteEntitiesParams, DeleteEntitiesResponseResult>;
+	ReadCommunity: Service<ReadCommunityParams, ReadCommunityResponseResult>;
+	ReadCommunities: Service<ReadCommunitiesParams, ReadCommunitiesResponseResult>;
+	UpdateCommunitySettings: Service<
+		UpdateCommunitySettingsParams,
+		UpdateCommunitySettingsResponseResult
+	>;
+	DeleteCommunity: Service<DeleteCommunityParams, DeleteCommunityResponseResult>;
+	ReadSpace: Service<ReadSpaceParams, ReadSpaceResponseResult>;
+	ReadSpaces: Service<ReadSpacesParams, ReadSpacesResponseResult>;
+	ReadEntities: Service<ReadEntitiesParams, ReadEntitiesResponseResult>;
+	ReadEntitiesPaginated: Service<ReadEntitiesPaginatedParams, ReadEntitiesPaginatedResponseResult>;
+	UpdateSpace: Service<UpdateSpaceParams, UpdateSpaceResponseResult>;
+	DeleteSpace: Service<DeleteSpaceParams, DeleteSpaceResponseResult>;
+	CreateTie: Service<CreateTieParams, CreateTieResponseResult>;
+	ReadTies: Service<ReadTiesParams, ReadTiesResponseResult>;
+	DeleteTie: Service<DeleteTieParams, DeleteTieResponseResult>;
 }
 
 // generated by src/lib/app/eventTypes.gen.ts
