+++ src/lib/vocab/persona/personaServices.ts
@@ -1,13 +1,7 @@
 import {blue, gray} from 'kleur/colors';
 import {Logger} from '@feltcoop/felt/util/log.js';
 
-import type {Service} from '$lib/server/service';
-import type {
-	CreateAccountPersonaParams,
-	CreateAccountPersonaResponseResult,
-	ReadPersonaParams,
-	ReadPersonaResponseResult,
-} from '$lib/app/eventTypes';
+import type {ServicesByName} from '$lib/app/eventTypes';
 import {CreateAccountPersona, ReadPersona} from '$lib/vocab/persona/personaEvents';
 import {toDefaultCommunitySettings} from '$lib/vocab/community/community.schema';
 import {createDefaultSpaces} from '$lib/vocab/space/spaceServices';
@@ -15,10 +9,7 @@ import {createDefaultSpaces} from '$lib/vocab/space/spaceServices';
 const log = new Logger(gray('[') + blue('personaServices') + gray(']'));
 
 //Creates a new persona
-export const createAccountPersonaService: Service<
-	CreateAccountPersonaParams,
-	CreateAccountPersonaResponseResult
-> = {
+export const createAccountPersonaService: ServicesByName['CreateAccountPersona'] = {
 	event: CreateAccountPersona,
 	// TODO verify the `account_id` has permission to modify this persona
 	// TODO add `actor_id` and verify it's one of the `account_id`'s personas
@@ -92,7 +83,7 @@ export const createAccountPersonaService: Service<
 };
 
 //Returns a single persona object
-export const readPersonaService: Service<ReadPersonaParams, ReadPersonaResponseResult> = {
+export const readPersonaService: ServicesByName['ReadPersona'] = {
 	event: ReadPersona,
 	perform: async ({repos, params}) => {
 		log.trace('[ReadPersona] persona', params.persona_id);
